refactor(backend): ServiceError 파일 분리
